fixing a lot of code and cleaning up repo
