+++ .gitignore
@@ -1,2 +1,3 @@
 **.vscode/**
 **/.ipynb_checkpoints/**
+g_sheet_creds.json
