--- .ipynb_checkpoints/README-checkpoint.md
@@ -1,26 +1,0 @@
-## NFL Fantasy Model & Application 
-
-### Build Process 
-    1. web scrape the data
-    2. clean data for writing to gdrive
-    3. drop into google sheets (defacto DB)
-    4. feature engineer 
-    5. model for player score 
-    6. develop web app for automation and dashboarding 
-    7. build optimazation process for line-up generation 
-
-### Process Status 
-    1. functional framework developed. Code optimization next step. 
-    2. cleaning process complete - NFL.com is the source of player data 
-    3. ability to write and scrape at scale - complete and functionalized 
-    4. player panel data ready for modeling. 
-        * looking to add data on opp defensive rank, active status and depth chart status 
-    5.  
-    6. WIP - current fuctionality:
-        1. Display output table 
-        2. upload & download CSV output 
-    7. Not developed 
-        * will require list of player salaries 
-        * predictions and variance 
-        * ability to input line-up parameters and restrictions
-
