--- 02_feature-engineering-data-prep/exp_smoothing.py
@@ -1,2 +1,0 @@
-def exp_smooth(df, smooth_col, alpha, lag):
-     series = lambda x: x.rolling(lag, 1)
