+++ requirements.txt
@@ -1,0 +1,174 @@
+alabaster==0.7.12
+argh==0.26.2
+arviz==0.9.0
+asn1crypto==1.2.0
+astroid==2.3.2
+atomicwrites==1.3.0
+attrs==19.3.0
+autopep8==1.4.4
+Babel==2.7.0
+backcall==0.1.0
+bcrypt==3.1.7
+beautifulsoup4==4.8.1
+bleach==3.1.0
+cachetools==3.1.1
+certifi==2020.6.20
+cloudpickle==1.2.2
+colorama==0.4.1
+conda
+conda-package-handling
+contextvars==2.4
+cryptography==2.3.1
+cycler==0.10.0
+dataclasses==0.7
+decorator==4.4.1
+defusedxml==0.6.0
+df2gspread==1.0.4
+diff-match-patch==20181111
+docutils==0.15.2
+entrypoints==0.3
+fastprogress==1.0.0
+flake8==3.7.9
+future==0.18.2
+google-api-python-client==1.6.7
+google-auth==1.7.1
+google-auth-httplib2==0.0.3
+google-auth-oauthlib==0.4.1
+gspread==3.1.0
+h5py==2.10.0
+helpdev==0.7.1
+httplib2==0.14.0
+idna==2.8
+imagesize==1.1.0
+immutables==0.14
+importlib-metadata==0.23
+intervaltree==3.0.2
+ipykernel==5.1.3
+ipython==7.9.0
+ipython-genutils==0.2.0
+ipywidgets==7.5.1
+isort==4.3.21
+jedi==0.14.1
+Jinja2==2.10.3
+joblib==0.13.2
+json5==0.8.5
+jsonschema==3.1.1
+jupyter==1.0.0
+jupyter-client==5.3.4
+jupyter-console==6.0.0
+jupyter-contrib-core==0.3.3
+jupyter-contrib-nbextensions==0.5.1
+jupyter-core==4.6.1
+jupyter-highlight-selected-word==0.2.0
+jupyter-latex-envs==1.4.6
+jupyter-nbextensions-configurator==0.4.1
+keyring==18.0.0
+kiwisolver==1.1.0
+lazy-object-proxy==1.4.2
+lxml==4.4.1
+Mako==1.1.3
+MarkupSafe==1.1.1
+matplotlib==3.3.0
+mccabe==0.6.1
+menuinst==1.4.16
+mistune==0.8.4
+mkl-fft==1.0.15
+mkl-random==1.1.0
+mkl-service==2.3.0
+more-itertools==7.2.0
+nbconvert==5.6.1
+nbformat==4.4.0
+netCDF4==1.5.3
+notebook==6.0.2
+numpy==1.17.3
+numpydoc==0.9.1
+oauth2client==4.1.3
+oauthlib==3.1.0
+packaging==19.2
+pandas==0.25.2
+pandocfilters==1.4.2
+paramiko==2.4.2
+parso==0.5.2
+pathtools==0.1.2
+patsy==0.5.1
+pexpect==4.7.0
+pickleshare==0.7.5
+Pillow==7.2.0
+pluggy==0.13.1
+progressbar2==3.47.0
+prometheus-client==0.7.1
+prompt-toolkit==2.0.10
+psutil==5.6.3
+ptyprocess==0.6.0
+pyasn1==0.4.8
+pyasn1-modules==0.2.7
+pycodestyle==2.5.0
+pycosat==0.6.3
+pycparser==2.19
+pydantic==1.3
+pydocstyle==4.0.1
+pyflakes==2.1.1
+pygam==0.8.0
+pygdrive3==0.6.0
+Pygments==2.4.2
+pygpu==0.7.6
+pygsheets==2.0.2
+pylint==2.4.3
+pymc3==3.9.3
+Pympler==0.7
+PyNaCl==1.3.0
+pyOpenSSL==19.0.0
+pyparsing==2.4.2
+PyQt5==5.12.3
+PyQt5-sip==12.8.1
+PyQtWebEngine==5.12.1
+pyreadline==2.1
+pyrsistent==0.15.4
+PySocks==1.7.1
+python-dateutil==2.8.0
+python-jsonrpc-server==0.3.2
+python-language-server==0.31.2
+python-utils==2.3.0
+pytz==2019.3
+pywin32==223
+pywinpty==0.5.5
+PyYAML==5.2
+pyzmq==18.1.0
+QDarkStyle==2.7
+QtAwesome==0.6.0
+qtconsole==4.6.0
+QtPy==1.9.0
+requests==2.22.0
+requests-oauthlib==1.3.0
+rope==0.14.0
+rsa==4.0
+Rtree==0.8.3
+ruamel-yaml==0.15.46
+scikit-learn==0.21.3
+scipy==1.3.1
+seaborn==0.10.1
+Send2Trash==1.5.0
+six==1.12.0
+snowballstemmer==2.0.0
+sortedcontainers==2.1.0
+soupsieve==1.9.3
+statsmodels==0.10.1
+terminado==0.8.2
+tqdm==4.36.1
+traitlets==4.3.3
+typed-ast==1.4.0
+typing-extensions==3.7.4.3
+ujson==1.35
+uritemplate==3.0.0
+urllib3==1.24.2
+watchdog==0.9.0
+wcwidth==0.1.7
+webencodings==0.5.1
+widgetsnbextension==3.5.1
+win-inet-pton==1.1.0
+wincertstore==0.2
+wrapt==1.11.2
+xarray==0.15.1
+xgboost==0.90
+yapf==0.28.0
+zipp==0.6.0
